moving finance classes in a package
