+++ src/main/java/io/mosfet/lm/exercise/products/Product.java
@@ -1,5 +1,7 @@
 package io.mosfet.lm.exercise.products;
 
+import io.mosfet.lm.exercise.cash.Money;
+
 public interface Product {
     Money getCost();
 }
